+++ CHANGES.rst
@@ -6,7 +6,7 @@
 * Added SQLAlchemy column annotations
 * Reorganised ``coaster.utils`` and ``coaster.sqlalchemy`` into sub-packages
 * ``LabeledEnum`` now supports grouped values if declared as a set
-* New: ``coaster.sqlalchemy.StateProperty`` adds state management to models
+* New: ``coaster.sqlalchemy.StateManager`` adds state management to models
 * New: ``coaster.utils.AttributeDict`` provides attribute access to a dict